+++ stroom-core-server/src/main/java/stroom/entity/server/EntityMarshaller.java
@@ -16,7 +16,6 @@
 
 package stroom.entity.server;
 
-import javassist.Modifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MarkerFactory;
@@ -25,9 +24,6 @@ import stroom.entity.shared.BaseEntity;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
-import java.lang.reflect.Array;
-import java.lang.reflect.Field;
-import java.util.Collection;
 
 public abstract class EntityMarshaller<E extends BaseEntity, O> implements Marshaller<E, O> {
     private static final Logger LOGGER = LoggerFactory.getLogger(EntityMarshaller.class);