+++ stroom-util-shared/src/main/java/stroom/util/shared/RandomId.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package stroom.util.client;
+package stroom.util.shared;
 
 public class RandomId {
     private static char[] CHARS = new char[]{'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D',